Use host's database

+++ infra/deployment/script/release_port.sh
@@ -1,4 +1,7 @@
 PORT=$1
 
-IFS=' ' read -r -a mgwarr <<< "$(netstat -aon | findstr $PORT)"
-taskkill //PID "$(sed -e 's/^[[:space:]]*//' <<<"${mgwarr[4]}")" //f
+if [[ $PORT != 5432 ]]; then
+  echo "Releasing port $PORT"
+  IFS=' ' read -r -a mgwarr <<< "$(netstat -aon | findstr $PORT)"
+  taskkill //PID "$(sed -e 's/^[[:space:]]*//' <<<"${mgwarr[4]}")" //f
+fi
